requirements.txt
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
Django==2.2
django-allauth==0.39.0
django-cors-headers==3.1.1
django-rest-auth==0.9.5
djangorestframework==3.10.3
idna==2.8
oauthlib==3.1.0
psycopg2-binary==2.8.3
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
urllib3==1.25.6
flake8>=3.6.0,<3.7.0
psycopg2>=2.7.5,<2.8.0
Pillow>=5.3.0,<5.4.0
pytest-django==3.5.1
